Add pipfile, gitignore, test, and use subprocess
The pipfile would allow adding any dependencies easier, and I had it
since I wanted to use a virtual env (pipenv shell) which automatically
created it. It also makes it explicit what python version this is
targetting.

The test checks that the public ip is correct by looking at another
website, using the built-in unittest library.

Subprocess is recommended to be used over os.popen in the python docs.
Plus, popen was returning an object. The stderr parameter is used
because curl is outputting its progress bar into stderr for some reason
so we set it to DEVNULL to not show it.

__init__.py was also added to make ipmonitor a proper python package (we
can import from it, which is what the test needs to do).

Furthermore, editor files were ignored and a typo in the readme was fixed.However, the project name itself also has a typo (maintenace => maintenance)