Turn into cli, use poetry, pre-commit, and add local + public ip
It is now a cli with 'cm' as the entry point. The project is also now managed with pyproject + poetry.

Running poetry install + poetry shell should get you setup and into the virtualenv.

Running the tests is as easy as poetry run task test. This is because taskipy is installed so 'test' is a task that is defined in pyproject.toml.

pre-commit install will install the git pre-commit hooks to auto format the files.

If you're installing this on the cluster, use poetry install --no-dev. click is the only other dependency other than python itself then.

bumpversion is used for releases. bump2version [major, minor, patch] automatically creates a git commit with the versions all updated and a tag attached to the git commit. Then you can just push up with git push && git push --tags.